Move character and item classes into folders
