+++ src/worldofzuul/Room.java
@@ -4,6 +4,8 @@ import java.util.ArrayList;
 import java.util.Set;
 import java.util.HashMap;
 import java.util.Iterator;
+import worldofzuul.item.Item;
+import worldofzuul.item.Key;
 
 
 /**
