+++ src/worldofzuul/item/Item.java
@@ -3,7 +3,8 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package worldofzuul;
+package worldofzuul.item;
+
 
 /**
  *
@@ -18,24 +19,21 @@ public class Item {
      * Creating a private boolean expression which by default is true
      */
     private boolean moveable = true;
-    
+
     /**
      * Creating a constructor for the class Item with 1 parameter
-     * @param name 
+     * @param name
      */
     public Item(String name) {
         this.name = name;
     }
     /**
-     * Creating a constructor for the class Item with 2 parameteres 
+     * Creating a constructor for the class Item with 2 parameteres
      * @param name
-     * @param moveable 
+     * @param moveable
      */
     public Item(String name, boolean moveable) {
         this.name = name;
         this.moveable = moveable;
     }
 }
-
-
-
